+++ packages/sequencer/src/protocol/production/sequencing/TransactionExecutionService.ts
@@ -1,8 +1,6 @@
 import { container, inject, injectable, Lifecycle, scoped } from "tsyringe";
 import {
   BlockProverExecutionData,
-  BlockProverState,
-  DefaultProvableHashList,
   NetworkState,
   Protocol,
   ProtocolModulesRecord,
@@ -10,24 +8,12 @@ import {
   RuntimeMethodExecutionContext,
   RuntimeMethodExecutionData,
   RuntimeProvableMethodExecutionResult,
-  RuntimeTransaction,
-  StateTransition,
-  ProvableBlockHook,
-  BlockHashMerkleTree,
   StateServiceProvider,
   MandatoryProtocolModulesRecord,
-  BlockHashTreeEntry,
-  MinaActions,
-  MinaActionsHashList,
   reduceStateTransitions,
 } from "@proto-kit/protocol";
-import { Bool, Field, Poseidon } from "o1js";
-import {
-  AreProofsEnabled,
-  log,
-  RollupMerkleTree,
-  mapSequential,
-} from "@proto-kit/common";
+import { Field } from "o1js";
+import { AreProofsEnabled, log, mapSequential } from "@proto-kit/common";
 import {
   MethodParameterEncoder,
   Runtime,
@@ -37,17 +23,8 @@ import {
 
 import { PendingTransaction } from "../../../mempool/PendingTransaction";
 import { CachedStateService } from "../../../state/state/CachedStateService";
-import { distinctByString } from "../../../helpers/utils";
-import { CachedMerkleTreeStore } from "../../../state/merkle/CachedMerkleTreeStore";
-import { AsyncMerkleTreeStore } from "../../../state/async/AsyncMerkleTreeStore";
-import {
-  TransactionExecutionResult,
-  Block,
-  BlockResult,
-  BlockWithResult,
-} from "../../../storage/model/Block";
+import { TransactionExecutionResult } from "../../../storage/model/Block";
 import { UntypedStateTransition } from "../helpers/UntypedStateTransition";
-import type { StateRecord } from "../BatchProducerModule";
 
 const errors = {
   methodIdNotFound: (methodId: string) =>
@@ -61,43 +38,16 @@ export type SomeRuntimeMethod = (...args: unknown[]) => Promise<unknown>;
 export class TransactionExecutionService {
   private readonly transactionHooks: ProvableTransactionHook<unknown>[];
 
-  private readonly blockHooks: ProvableBlockHook<unknown>[];
-
   public constructor(
     @inject("Runtime") private readonly runtime: Runtime<RuntimeModulesRecord>,
     @inject("Protocol")
-    private readonly protocol: Protocol<
-      MandatoryProtocolModulesRecord & ProtocolModulesRecord
-    >,
-    private readonly executionContext: RuntimeMethodExecutionContext,
+    protocol: Protocol<MandatoryProtocolModulesRecord & ProtocolModulesRecord>,
     // Coming in from the appchain scope (accessible by protocol & runtime)
     @inject("StateServiceProvider")
     private readonly stateServiceProvider: StateServiceProvider
   ) {
     this.transactionHooks = protocol.dependencyContainer.resolveAll(
       "ProvableTransactionHook"
-    );
-    this.blockHooks =
-      protocol.dependencyContainer.resolveAll("ProvableBlockHook");
-  }
-
-  private allKeys(stateTransitions: StateTransition<unknown>[]): Field[] {
-    // We have to do the distinct with strings because
-    // array.indexOf() doesn't work with fields
-    return stateTransitions.map((st) => st.path).filter(distinctByString);
-  }
-
-  private collectStateDiff(
-    stateTransitions: UntypedStateTransition[]
-  ): StateRecord {
-    return stateTransitions.reduce<Record<string, Field[] | undefined>>(
-      (state, st) => {
-        if (st.toValue.isSome.toBoolean()) {
-          state[st.path.toString()] = st.toValue.value;
-        }
-        return state;
-      },
-      {}
     );
   }
 
@@ -217,212 +167,7 @@ export class TransactionExecutionService {
     );
   }
 
-  /**
-   * Main entry point for creating a unproven block with everything
-   * attached that is needed for tracing
-   */
-  public async createBlock(
-    stateService: CachedStateService,
-    transactions: PendingTransaction[],
-    lastBlockWithResult: BlockWithResult,
-    allowEmptyBlocks: boolean
-  ): Promise<Block | undefined> {
-    const lastResult = lastBlockWithResult.result;
-    const lastBlock = lastBlockWithResult.block;
-    const executionResults: TransactionExecutionResult[] = [];
-
-    const transactionsHashList = new DefaultProvableHashList(Field);
-    const eternalTransactionsHashList = new DefaultProvableHashList(
-      Field,
-      Field(lastBlock.toEternalTransactionsHash)
-    );
-
-    const incomingMessagesList = new MinaActionsHashList(
-      Field(lastBlock.toMessagesHash)
-    );
-
-    // Get used networkState by executing beforeBlock() hooks
-    const networkState = await this.blockHooks.reduce<Promise<NetworkState>>(
-      async (reduceNetworkState, hook) =>
-        await hook.beforeBlock(await reduceNetworkState, {
-          blockHashRoot: Field(lastResult.blockHashRoot),
-          eternalTransactionsHash: lastBlock.toEternalTransactionsHash,
-          stateRoot: Field(lastResult.stateRoot),
-          transactionsHash: Field(0),
-          networkStateHash: lastResult.afterNetworkState.hash(),
-          incomingMessagesHash: lastBlock.toMessagesHash,
-        }),
-      Promise.resolve(lastResult.afterNetworkState)
-    );
-
-    for (const [, tx] of transactions.entries()) {
-      try {
-        // Create execution trace
-        // eslint-disable-next-line no-await-in-loop
-        const executionTrace = await this.createExecutionTrace(
-          stateService,
-          tx,
-          networkState
-        );
-
-        // Push result to results and transaction onto bundle-hash
-        executionResults.push(executionTrace);
-        if (!tx.isMessage) {
-          transactionsHashList.push(tx.hash());
-          eternalTransactionsHashList.push(tx.hash());
-        } else {
-          const actionHash = MinaActions.actionHash(
-            tx.toRuntimeTransaction().hashData()
-          );
-
-          incomingMessagesList.push(actionHash);
-        }
-      } catch (error) {
-        if (error instanceof Error) {
-          log.error("Error in inclusion of tx, skipping", error);
-        }
-      }
-    }
-
-    const previousBlockHash =
-      lastResult.blockHash === 0n ? undefined : Field(lastResult.blockHash);
-
-    if (executionResults.length === 0 && !allowEmptyBlocks) {
-      log.info(
-        "After sequencing, block has no sequencable transactions left, skipping block"
-      );
-      return undefined;
-    }
-
-    const block: Omit<Block, "hash"> = {
-      transactions: executionResults,
-      transactionsHash: transactionsHashList.commitment,
-      fromEternalTransactionsHash: lastBlock.toEternalTransactionsHash,
-      toEternalTransactionsHash: eternalTransactionsHashList.commitment,
-      height:
-        lastBlock.hash.toBigInt() !== 0n ? lastBlock.height.add(1) : Field(0),
-      fromBlockHashRoot: Field(lastResult.blockHashRoot),
-      fromMessagesHash: lastBlock.toMessagesHash,
-      toMessagesHash: incomingMessagesList.commitment,
-      previousBlockHash,
-
-      networkState: {
-        before: new NetworkState(lastResult.afterNetworkState),
-        during: networkState,
-      },
-    };
-
-    const hash = Block.hash(block);
-
-    return {
-      ...block,
-      hash,
-    };
-  }
-
-  public async generateMetadataForNextBlock(
-    block: Block,
-    merkleTreeStore: AsyncMerkleTreeStore,
-    blockHashTreeStore: AsyncMerkleTreeStore,
-    modifyTreeStore = true
-  ): Promise<BlockResult> {
-    // Flatten diff list into a single diff by applying them over each other
-    const combinedDiff = block.transactions
-      .map((tx) => {
-        const transitions = tx.protocolTransitions.concat(
-          tx.status.toBoolean() ? tx.stateTransitions : []
-        );
-        return this.collectStateDiff(transitions);
-      })
-      .reduce<StateRecord>((accumulator, diff) => {
-        // accumulator properties will be overwritten by diff's values
-        return Object.assign(accumulator, diff);
-      }, {});
-
-    const inMemoryStore = new CachedMerkleTreeStore(merkleTreeStore);
-    const tree = new RollupMerkleTree(inMemoryStore);
-    const blockHashInMemoryStore = new CachedMerkleTreeStore(
-      blockHashTreeStore
-    );
-    const blockHashTree = new BlockHashMerkleTree(blockHashInMemoryStore);
-
-    await inMemoryStore.preloadKeys(Object.keys(combinedDiff).map(BigInt));
-
-    // In case the diff is empty, we preload key 0 in order to
-    // retrieve the root, which we need later
-    if (Object.keys(combinedDiff).length === 0) {
-      await inMemoryStore.preloadKey(0n);
-    }
-
-    // TODO This can be optimized a lot (we are only interested in the root at this step)
-    await blockHashInMemoryStore.preloadKey(block.height.toBigInt());
-
-    Object.entries(combinedDiff).forEach(([key, state]) => {
-      const treeValue = state !== undefined ? Poseidon.hash(state) : Field(0);
-      tree.setLeaf(BigInt(key), treeValue);
-    });
-
-    const stateRoot = tree.getRoot();
-    const fromBlockHashRoot = blockHashTree.getRoot();
-
-    const state: BlockProverState = {
-      stateRoot,
-      transactionsHash: block.transactionsHash,
-      networkStateHash: block.networkState.during.hash(),
-      eternalTransactionsHash: block.toEternalTransactionsHash,
-      blockHashRoot: fromBlockHashRoot,
-      incomingMessagesHash: block.toMessagesHash,
-    };
-
-    // TODO Set StateProvider for @state access to state
-    this.executionContext.clear();
-    this.executionContext.setup({
-      networkState: block.networkState.during,
-      transaction: RuntimeTransaction.dummyTransaction(),
-    });
-
-    const resultingNetworkState = await this.blockHooks.reduce<
-      Promise<NetworkState>
-    >(
-      async (networkState, hook) =>
-        await hook.afterBlock(await networkState, state),
-      Promise.resolve(block.networkState.during)
-    );
-
-    const { stateTransitions } = this.executionContext.result;
-    this.executionContext.clear();
-    const reducedStateTransitions = reduceStateTransitions(stateTransitions);
-
-    // Update the block hash tree with this block
-    blockHashTree.setLeaf(
-      block.height.toBigInt(),
-      new BlockHashTreeEntry({
-        blockHash: Poseidon.hash([block.height, state.transactionsHash]),
-        closed: Bool(true),
-      }).hash()
-    );
-    const blockHashWitness = blockHashTree.getWitness(block.height.toBigInt());
-    const newBlockHashRoot = blockHashTree.getRoot();
-    await blockHashInMemoryStore.mergeIntoParent();
-
-    if (modifyTreeStore) {
-      await inMemoryStore.mergeIntoParent();
-    }
-
-    return {
-      afterNetworkState: resultingNetworkState,
-      stateRoot: stateRoot.toBigInt(),
-      blockHashRoot: newBlockHashRoot.toBigInt(),
-      blockHashWitness,
-
-      blockStateTransitions: reducedStateTransitions.map((st) =>
-        UntypedStateTransition.fromStateTransition(st)
-      ),
-      blockHash: block.hash.toBigInt(),
-    };
-  }
-
-  private async createExecutionTrace(
+  public async createExecutionTrace(
     asyncStateService: CachedStateService,
     tx: PendingTransaction,
     networkState: NetworkState
